+++ rio_tiler/__init__.py
@@ -1,6 +1,6 @@
 """rio-tiler."""
 
-__version__ = "7.2.0"
+__version__ = "7.2.2"
 
 from . import (  # noqa
     colormap,